% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_bf}
\alias{calc_bf}
\title{Compute a Bayes factor}
\usage{
calc_bf(data, h1, h2, scale, nsamp = 1000)
}
\arguments{
\item{data}{A matrix. The dataset for which the BF must be computed}

\item{h1}{A constraint matrix defining H1.}

\item{h2}{A constraint matrix defining H2.}

\item{scale}{A number specifying the prior scale.}

\item{nsamp}{A number. The number of prior or posterior samples to determine the}
}
\value{
BF12, that is, the evidence for H1 relative to H2
}
\description{
Compute a Bayes factor
}
