% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{calc_fc}
\alias{calc_fc}
\title{Compute the complexity or fit for two hypotheses.}
\usage{
calc_fc(hyp, hyp2, means, sds, nsamp = 1000)
}
\arguments{
\item{hyp}{A constraint matrix defining H1.}

\item{means}{A vector of posterior or prior means}

\item{sds}{A vector or posterior or prior standard deviation}

\item{nsamp}{A number. The number of prior or posterior samples to determine the
fit and complexity}

\item{hyps2}{A constraint matrix defining H2 OR a character \code{'u'}
or \code{'c'} specifying an unconstrained or complement hypothesis}
}
\value{
A vector.
The proportion of posterior samples in agreement with H1 and with H2
}
\description{
Compute the complexity or fit for two hypotheses.
}
\examples{
# Complexity
priormeans <- c(0,0,0)
priorsds <- c(1,1,1)
h1 <- matrix(c(1,-1,0,0,1,-1), nrow= 2, byrow= TRUE)
h2 <- 'c'
complexity <- calc_fc(h1, h2, priormeans, priorsds)

# Fit
truemeans <- c(.4, .2, 0)
n <- 30
ngroup <- length(truemeans)
dat <- matrix(rnorm(n * ngroup, truemeans), ncol = ngroup, byrow = TRUE)
postmeans <- colMeans(dat)
postsds <- apply(dat, 2, sd)/sqrt(n)
fit <- calc_fc(h1, h2, postmeans, postsds)

}
