% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{eval_const}
\alias{eval_const}
\title{Evaluate a contraint matrix for a set of prior/posterior samples}
\usage{
eval_const(hyp, samples)
}
\arguments{
\item{hyp}{A constraint matrix defining a hypothesis.}

\item{samples}{A matrix. Prior or posterior samples, the number of columns
corresponds to the number of groups, the number of rows the number of samples}
}
\value{
A number between 0 and 1. The proportion of samples in which the
constraints are met.
}
\description{
Evaluate a contraint matrix for a set of prior/posterior samples
}
