% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{samp_bf}
\alias{samp_bf}
\title{Sample multiple datasets and compute the Bayes factor in each}
\usage{
samp_bf(datasets, n, ngroup, means, sds, h1, h2, comp)
}
\arguments{
\item{datasets}{A number. The number of datasets to simulate for each
sample size \code{n}}

\item{n}{A number. The group sample size to be used in data simulation}

\item{ngroup}{A number. The number of groups.}

\item{means}{A vector of expected population means (standardized).}

\item{sds}{A vector of expected population effect sizes.
Note, when standardized, this is a vector of 1s}

\item{h1}{A constraint matrix defining H1.}

\item{h2}{A constraint matrix defining H2.}

\item{comp}{A vector of the complexities of H1 and H2.
See \code{?BayesianPower::calc_fc}}
}
\value{
A vector of Bayes factors BF12 for each of the simulated datasets
}
\description{
Sample multiple datasets and compute the Bayes factor in each
}
