% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{samp_dist}
\alias{samp_dist}
\title{Sample from prior or posterior distribution}
\usage{
samp_dist(nsamp, means, sds)
}
\arguments{
\item{nsamp}{A number. The number of prior or posterior samples to determine the
fit and complexity}

\item{means}{A vector. The prior or posterior means for each group}

\item{sds}{A number or a vector. The standard deviations for each group
If a number is used, the same prior or posterior standard deviation is
used for each group.}
}
\value{
A matrix of \code{nsamp} rows and as many columns as the
length of \code{means}. Posterior or prior samples
}
\description{
Sample from prior or posterior distribution
}
