% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{calc_fc}
\alias{calc_fc}
\title{Compute the complexity or fit for two hypotheses.}
\usage{
calc_fc(hyp, hyp2, means, sds, nsamp = 1000)
}
\arguments{
\item{hyp}{A constraint matrix defining H1.}

\item{hyp2}{A constraint matrix defining H2 OR a character \code{'u'}
or \code{'c'} specifying an unconstrained or complement hypothesis}

\item{means}{A vector of posterior or prior means}

\item{sds}{A vector or posterior or prior standard deviation}

\item{nsamp}{A number. The number of prior or posterior samples to determine the
fit and complexity}
}
\value{
A vector.
The proportion of posterior samples in agreement with H1 and with H2
}
\description{
Compute the complexity or fit for two hypotheses.
}
